Fix the pause menu fade-in (#176858341)
Fixes a rare occurrence of the fade-in not working. When the fade-out
animation has already been turned on and the fade-in is enabled – so
that the fade-out's disable event occurs during the transition between
the fade animations – the overlay game object is no longer being
disabled when it's not needed.